Update references to moved files
Update references to
* local_global_conversion.py
* drone_odometry_local.py
